make new component for demographics chart
