+++ src/dataset-builder/TestConstants.ts
@@ -176,77 +176,6 @@ export const dummyGetConceptForId = (id: number): Concept => {
   return _.find({ id }, dummyConcepts)!;
 };
 
-interface ChartLabel {
-  short: string;
-  long?: string;
-}
-interface ChartSeries {
-  name?: string;
-  data: number[];
-}
-
-export interface CohortDemographics {
-  categories: ChartLabel[];
-  series: ChartSeries[];
-  title: string;
-  yTitle: string;
-  height: string;
-  legendEnabled: boolean;
-  showSeriesName: boolean;
-}
-
-export function generateCohortAgeData(series) {
-  return {
-    categories: [
-      { short: 'Female' },
-      { short: 'Male' },
-      { short: 'Other', long: 'Nonbinary, 2 Spirit, Genderqueer, etc.' },
-    ],
-    series,
-    title: 'Gender identity and current age',
-    yTitle: 'AVERAGE AGE',
-    height: '250rem',
-    legendEnabled: false,
-    showSeriesName: false,
-  };
-}
-
-export function generateRandomCohortAgeSeries() {
-  return [{ data: generateRandomNumbers(3, 90) }];
-}
-export function generateRandomCohortAgeData() {
-  return generateCohortAgeData(generateRandomCohortAgeSeries());
-}
-
-export function generateCohortDemographicData(series) {
-  return {
-    categories: [
-      { short: 'Female' },
-      { short: 'Female 18-44' },
-      { short: 'Female 45-64' },
-      { short: 'Female 65+' },
-      { short: 'Male' },
-      { short: 'Male 18-44' },
-      { short: 'Male 45-64' },
-      { short: 'Male 65+' },
-      { short: 'Other', long: 'Nonbinary, 2 Spirit, Genderqueer, etc.' },
-      { short: 'Other 18-44', long: 'Nonbinary, 2 Spirit, Genderqueer, etc. 18-44' },
-      { short: 'Other 45-64', long: 'Nonbinary, 2 Spirit, Genderqueer, etc. 45-64' },
-      { short: 'Other 65+', long: 'Nonbinary, 2 Spirit, Genderqueer, etc. 65+' },
-    ],
-    series,
-    title: 'Gender identity, current age, and race',
-    yTitle: 'OVERALL PERCENTAGE',
-    height: '500rem',
-    legendEnabled: true,
-    showSeriesName: true,
-  };
-}
-
-export function generateRandomCohortDemographicData() {
-  return generateCohortDemographicData(generateDemographicSeries());
-}
-
 export function generateRandomNumbers(numNumbers: number, max: number) {
   const randomNumbers: number[] = [];
   for (let i = 0; i < numNumbers; i++) {
@@ -266,37 +195,4 @@ export function generateRandomNumbersThatAddUpTo(total: number, numNumbers: numb
   }
   randomNumbers.push(remaining);
   return _.shuffle(randomNumbers);
-}
-
-// series and data will always be the same length
-export function addToSeriesData(series: ChartSeries[], data: number[]) {
-  if (series.length !== data.length) throw new Error('series and data must be the same length');
-  for (let i = 0; i < series.length; i++) {
-    series[i].data.push(data[i]);
-  }
-}
-
-export function generateDemographicSeries(): ChartSeries[] {
-  const series: ChartSeries[] = [
-    { name: 'Asian', data: [] },
-    { name: 'Black', data: [] },
-    { name: 'White', data: [] },
-    { name: 'Native American', data: [] },
-    { name: 'Pacific Islander', data: [] },
-  ];
-  // for each of the three gender identity totals,
-  const genderTotals = generateRandomNumbersThatAddUpTo(100, 3);
-  for (const genderTotal of genderTotals) {
-    // generate the race breakdown for the gender totals
-    const genderTotalRaceBreakDown = generateRandomNumbersThatAddUpTo(genderTotal, 5);
-    addToSeriesData(series, genderTotalRaceBreakDown);
-    // get the three age group breakdowns
-    const ageGroupGenderTotal = generateRandomNumbersThatAddUpTo(genderTotal, 3);
-    // get race breakdowns for each of the gender age groups
-    for (const genderAgeGroupTotal of ageGroupGenderTotal) {
-      const genderAgeGroupRaceBreakdown = generateRandomNumbersThatAddUpTo(genderAgeGroupTotal, 5);
-      addToSeriesData(series, genderAgeGroupRaceBreakdown);
-    }
-  }
-  return series;
 }
